feat: 회원가입 시 이미지 추가

+++ src/main/kotlin/udumeoli/tripphoto/user/dto/UserPayload.kt
@@ -5,7 +5,7 @@ import udumeoli.tripphoto.user.entity.ServiceUser
 data class UserPayload(
     val id: Long,
     val nickname: String,
-    val profileImage: Int,
+    val profileImage: Long,
 )
 
 fun ServiceUser.toPayload(): UserPayload =
